backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
openai>=1.0.0
yfinance==0.2.50
pandas==2.2.3
ta==0.11.0
requests==2.32.3
feedparser==6.0.11
apscheduler==3.10.4
python-dotenv==1.0.1
httpx==0.28.1
